Removed redundant files. Organized models in namespaces. Changed jQuery paging.
