+++ Core.DomainModel/Core.DomainModel.csproj
@@ -58,12 +58,15 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ApplicationUser.cs" />
+    <Compile Include="Course.cs" />
     <Compile Include="ICreatedOn.cs" />
-    <Compile Include="IIdentifier.cs" />
+    <Compile Include="IEntity.cs" />
     <Compile Include="IModifiedOn.cs" />
     <Compile Include="Mail\EmailModel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Models.cs" />
+    <Compile Include="ClassRoom.cs" />
+    <Compile Include="Student.cs" />
+    <Compile Include="Teacher.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
