+++ Core.DomainModel/Course.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Core.DomainModel
+{
+    public class Course : IEntity
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public virtual ICollection<Student> Students { get; set; }
+    }
+}
