+++ Core.DomainModel/IEntity.cs
