+++ Core.DomainModel/Teacher.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+
+namespace Core.DomainModel
+{
+    public class Teacher : IEntity
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public virtual ICollection<Course> Courses { get; set; }
+    }
+}
