+++ Infrastructure.DataAccess/Infrastructure.DataAccess.csproj
@@ -60,11 +60,13 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ChangeSampleSeedInitializer.cs" />
     <Compile Include="GenericRepository.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SampleContext.cs" />
-    <Compile Include="SampleSeedInitializer.cs" />
-    <Compile Include="Seeding\Seeding.cs" />
+    <Compile Include="CreateSampleSeedInitializer.cs" />
+    <Compile Include="SeedHelper.cs" />
+    <Compile Include="Seeding\SeedingHelper.cs" />
     <Compile Include="UnitOfWork.cs" />
   </ItemGroup>
   <ItemGroup>
