+++ Presentation.Web/App_Start/IdentityConfig.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Net.Mail;
-using System.Net.Mime;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using Core.DomainModel;
@@ -89,56 +87,6 @@ namespace Presentation.Web.App_Start
         public static ApplicationSignInManager Create(IdentityFactoryOptions<ApplicationSignInManager> options, IOwinContext context)
         {
             return new ApplicationSignInManager(context.GetUserManager<ApplicationUserManager>(), context.Authentication);
-        }
-    }
-
-    public class EmailService : IIdentityMessageService
-    {
-        private readonly SmtpClient _smtpClient;
-
-        public EmailService(SmtpClient smtpClient)
-        {
-            _smtpClient = smtpClient;
-        }
-
-        public Task SendAsync(IdentityMessage message)
-        {
-            var mailMessage = new MailMessage()
-            {
-                Subject = message.Subject,
-                IsBodyHtml = true,
-                Body = message.Body
-            };
-            
-            mailMessage.To.Add(message.Destination);
-
-            // alternate view, for clients that can't view HTML
-            var mimeType = new ContentType("text/html");
-            var alternate = AlternateView.CreateAlternateViewFromString(message.Body, mimeType);
-            mailMessage.AlternateViews.Add(alternate);
-            
-            try
-            {
-                _smtpClient.SendAsync(mailMessage, null);
-            }
-            catch (Exception e)
-            {
-                // Handle errors here.
-                // For example someone typed in a wrong email address.
-                // Right now this check is redundant, as it throws on the next line.
-                throw;
-            }
-
-            return Task.FromResult(0);
-        }
-    }
-
-    public class SmsService : IIdentityMessageService
-    {
-        public Task SendAsync(IdentityMessage message)
-        {
-            // Plug in your SMS service here to send a text message.
-            return Task.FromResult(0);
         }
     }
 }
