+++ Presentation.Web/App_Start/MappingConfig.cs
@@ -2,6 +2,8 @@
 using Core.DomainModel;
 using Presentation.Web.App_Start;
 using Presentation.Web.Models;
+using Presentation.Web.Models.Account;
+using Presentation.Web.Models.Student;
 
 [assembly: WebActivatorEx.PreApplicationStartMethod(typeof(MappingConfig), "Start")]
 
@@ -19,8 +21,8 @@ namespace Presentation.Web.App_Start
             // Write your AutoMapper configurations here.
 
             // ViewModel Mappings
-            Mapper.CreateMap<StudentViewModel, Student>().ReverseMap();
-            Mapper.CreateMap<RegisterViewModelExample, ApplicationUser>().ReverseMap();
+            Mapper.CreateMap<NewStudentViewModel, Student>().ReverseMap();
+            Mapper.CreateMap<RegisterViewModel, ApplicationUser>().ReverseMap();
         }
     }
 }
