+++ Presentation.Web/App_Start/SmsService.cs
@@ -1,0 +1,14 @@
+﻿using System.Threading.Tasks;
+using Microsoft.AspNet.Identity;
+
+namespace Presentation.Web.App_Start
+{
+    public class SmsService : IIdentityMessageService
+    {
+        public Task SendAsync(IdentityMessage message)
+        {
+            // Plug in your SMS service here to send a text message.
+            return Task.FromResult(0);
+        }
+    }
+}
