+++ Presentation.Web/MailTemplates/EmailTemplate.cshtml
@@ -1,4 +1,4 @@
-﻿@model Presentation.Web.Models.MailViewModel
+﻿@model Presentation.Web.Models.Student.MailViewModel
 <!DOCTYPE html>
 <html>
 <head>
