+++ Presentation.Web/Models/Account/ForgotPasswordViewModel.cs
@@ -1,0 +1,7 @@
+﻿namespace Presentation.Web.Models
+{
+    public class ForgotPasswordViewModel
+    {
+        public string Email { get; set; }
+    }
+}
