+++ Presentation.Web/Models/Account/RegisterViewModel.cs
@@ -1,0 +1,10 @@
+﻿namespace Presentation.Web.Models.Account
+{
+    public class RegisterViewModel
+    {
+        public string UserName { get; set; }
+        public string Email { get; set; }
+        public string Password { get; set; }
+        public string ConfirmPassword { get; set; }
+    }
+}
