+++ Presentation.Web/Models/Manage/IndexViewModel.cs
@@ -1,0 +1,7 @@
+﻿namespace Presentation.Web.Models.Manage
+{
+    public class IndexViewModel
+    {
+        public bool HasPassword { get; set; }
+    }
+}
