--- Presentation.Web/Models/ManageViewModelExample.cs
@@ -1,14 +1,0 @@
-﻿namespace Presentation.Web.Models
-{
-    public class ManageIndexViewModelExample
-    {
-        public bool HasPassword { get; set; }
-    }
-
-    public class ChangePasswordViewModelExample
-    {
-        public string OldPassword { get; set; }
-        public string NewPassword { get; set; }
-        public string ConfirmPassword { get; set; }
-    }
-}
