+++ Presentation.Web/Models/Student/MailViewModel.cs
@@ -1,0 +1,12 @@
+﻿namespace Presentation.Web.Models.Student
+{
+    public class MailViewModel
+    {
+        public string Subject { get; set; }
+        public string Email { get; set; }
+        public string Foo { get; set; }
+        public string Bar { get; set; }
+        public string Baz { get; set; }
+        public string Qux { get; set; }
+    }
+}
