+++ Presentation.Web/Models/Student/NewStudentViewModel.cs
@@ -1,0 +1,8 @@
+﻿namespace Presentation.Web.Models.Student
+{
+    public class NewStudentViewModel
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+    }
+}
