+++ Presentation.Web/Presentation.Web.csproj
@@ -180,26 +180,32 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="App_Start\BundleConfig.cs" />
+    <Compile Include="App_Start\EmailService.cs" />
     <Compile Include="App_Start\FilterConfig.cs" />
     <Compile Include="App_Start\IdentityConfig.cs" />
     <Compile Include="App_Start\MappingConfig.cs" />
     <Compile Include="App_Start\NinjectWebCommon.cs" />
     <Compile Include="App_Start\RouteConfig.cs" />
+    <Compile Include="App_Start\SmsService.cs" />
     <Compile Include="App_Start\Startup.Auth.cs" />
-    <Compile Include="Controllers\ExampleAccountController.cs" />
-    <Compile Include="Controllers\ExampleManageController.cs" />
-    <Compile Include="Controllers\ExampleStudentController.cs" />
+    <Compile Include="Controllers\AccountController.cs" />
+    <Compile Include="Controllers\ManageController.cs" />
+    <Compile Include="Controllers\StudentController.cs" />
     <Compile Include="Controllers\HomeController.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
     <Compile Include="Mail\IMailHandler.cs" />
     <Compile Include="Mail\MailHandler.cs" />
-    <Compile Include="Models\AccountViewModelsExample.cs" />
-    <Compile Include="Models\ManageViewModelExample.cs" />
-    <Compile Include="Models\ManageViewModels.cs" />
-    <Compile Include="Models\PagedData.cs" />
-    <Compile Include="Models\StudentViewModels.cs" />
+    <Compile Include="Models\Account\ForgotPasswordViewModel.cs" />
+    <Compile Include="Models\Account\LoginViewModel.cs" />
+    <Compile Include="Models\Student\IndexViewModel.cs" />
+    <Compile Include="Models\Manage\IndexViewModel.cs" />
+    <Compile Include="Models\Manage\ChangePasswordViewModel.cs" />
+    <Compile Include="Models\Student\NewStudentViewModel.cs" />
+    <Compile Include="Models\Student\PagedData.cs" />
+    <Compile Include="Models\Account\RegisterViewModel.cs" />
+    <Compile Include="Models\Student\MailViewModel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Startup.cs" />
     <Content Include="Content\bootstrap-theme.css" />
@@ -259,12 +265,12 @@
     <Content Include="Views\Shared\Lockout.cshtml" />
     <Content Include="Views\Shared\_LoginPartial.cshtml" />
     <Content Include="Views\Home\AuthorizeAction.cshtml" />
-    <Content Include="Views\ExampleAccount\Register.cshtml" />
-    <Content Include="Views\ExampleAccount\Login.cshtml" />
-    <Content Include="Views\ExampleAccount\ForgotPassword.cshtml" />
-    <Content Include="Views\ExampleManage\Index.cshtml" />
-    <Content Include="Views\ExampleManage\ChangePassword.cshtml" />
-    <Content Include="Views\Student\TestMail.cshtml" />
+    <Content Include="Views\Account\Register.cshtml" />
+    <Content Include="Views\Account\Login.cshtml" />
+    <Content Include="Views\Account\ForgotPassword.cshtml" />
+    <Content Include="Views\Manage\Index.cshtml" />
+    <Content Include="Views\Manage\ChangePassword.cshtml" />
+    <Content Include="Views\Student\Mail.cshtml" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="App_Data\" />
