+++ Presentation.Web/Views/Account/ForgotPassword.cshtml
@@ -1,11 +1,11 @@
-﻿@model Presentation.Web.Models.ForgotPasswordViewModelExample
+﻿@model Presentation.Web.Models.ForgotPasswordViewModel
 @{
     ViewBag.Title = "ForgotPassword";
 }
 
 <h2>ForgotPassword</h2>
 
-@using (Html.BeginForm("ForgotPassword", "ExampleAccount", FormMethod.Post, new {}))
+@using (Html.BeginForm("ForgotPassword", "Account", FormMethod.Post, new {}))
 {
     @Html.AntiForgeryToken()
     @Html.LabelFor(m => m.Email)
