+++ Presentation.Web/Views/Account/Register.cshtml
@@ -1,11 +1,11 @@
-﻿@model Presentation.Web.Models.RegisterViewModelExample
+﻿@model Presentation.Web.Models.Account.RegisterViewModel
 @{
     ViewBag.Title = "Register";
 }
 
 <h2>@ViewBag.Title</h2>
 
-@using (@Html.BeginForm("Register", "ExampleAccount", FormMethod.Post, new {})) 
+@using (@Html.BeginForm("Register", "Account", FormMethod.Post, new {})) 
 {
     @Html.AntiForgeryToken()
     @Html.LabelFor(m => m.UserName)
