+++ Presentation.Web/Views/Manage/ChangePassword.cshtml
@@ -1,11 +1,11 @@
-﻿@model Presentation.Web.Models.ChangePasswordViewModelExample
+﻿@model Presentation.Web.Models.Manage.ChangePasswordViewModel
 @{
     ViewBag.Title = "ChangePassword";
 }
 
 <h2>Change Password</h2>
 
-@using (Html.BeginForm("ChangePassword", "ExampleManage", FormMethod.Post, new {}))
+@using (Html.BeginForm("ChangePassword", "Manage", FormMethod.Post, new {}))
 {
     @Html.AntiForgeryToken()
