+++ Presentation.Web/Views/Manage/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model Presentation.Web.Models.ManageIndexViewModelExample
+﻿@model Presentation.Web.Models.Manage.IndexViewModel
 @{
     ViewBag.Title = "Index";
 }
