+++ Presentation.Web/Views/Student/Mail.cshtml
@@ -1,4 +1,4 @@
-﻿@model Presentation.Web.Models.MailViewModel
+﻿@model Presentation.Web.Models.Student.MailViewModel
 <h2>Submit Email</h2>
 
 @using (Html.BeginForm("SendMail", "Student", new { }, FormMethod.Post, new { }))
