+++ Presentation.Web/Views/Student/NewStudent.cshtml
@@ -1,4 +1,4 @@
-﻿@model Presentation.Web.Models.StudentViewModel
+﻿@model Presentation.Web.Models.Student.NewStudentViewModel
 @{
     ViewBag.Title = "NewStudent";
 }
